Support two new substitutions:
BIBTEX2WEB_NOTICE is replaced by
"This page was generated $timestamp by "
."<a href=\"http://pag.lcs.mit.edu/~mernst/software/#bibtex2web\">"
."bibtex2web</a>"
where $timestamp is the local time in ctime(3) format.

COPYRIGHT_NOTICE is replaced by the contents of the file specified by
the -copyright parameter.

Bug: this only works for files generated by bwconv directly, so it
doesn't work yet for the author index page (which uses
make-author-pages.pl) or the per-paper pages (which uses
htmlabstract-split.pl).