Add htmllist output format: generates a list of entry titles, each of which links to the abstract page for that entry, separated by <br>eaks. This is useful to generate a list of "recent publications" on a home page.  For example, see "Selected Publications" list on http://pmg.csail.mit.edu.
Example make rule:
index.html:
	${BWBIN}/bwconv.pl -format=bibtex,htmllist -outopts=limit:5\ abstract_dir:pubs -headfoot index-headfoot.html -copyright ../copyright -to $@ $(FILTER) ${BIBFILES}

The "limit" output option limits the list to the specified number of
entries (without this option, list is unlimited).  The "abstract_dir"
option specifies the relative directory containing the per-entry
abstract files (default "../pubs").

Bugs: generates too much space in the HTML (a newline is generated for
each entry, even if it's omitted!  Is there a way to tell the system to
stop processing entries?).